megacheck

+++ iohlp/wrap_test.go
@@ -45,7 +45,7 @@ culpa qui officia deserunt mollit anim id est laborum.
 			"Tisztelt Kobe! Mai nap folyamán kaptam önöktől egy levelet 2015.12.03.-an történt egy baleset és a levélbe csatoltak még egy résztvevői nyiltakozatott és kérik hogy 5 napon belül jutasam vissza önökhöz de én már másnap voltam a tatabányai kirendeltségbe és egy ugyan ilyet már kitöltöttek velünk kérdésem az hogy mi lett azzal a papíral vagy ez a levél érkezett meg késve ? Válaszát előre is köszönöm! TISZTELETTEL: Tromposch Tamás ",
 		},
 	} {
-		b, err := ioutil.ReadAll(WrappingReader(strings.NewReader(tc.in), tc.width))
+		b, err := ioutil.ReadAll(WrappingReader(strings.NewReader(tc.in), uint(tc.width)))
 		if err != nil {
 			t.Errorf("%d: %v", i, err)
 			continue
